Decompose grpc_server.go into smaller focused files
Split the 5225-line grpc_server.go into 12 focused files to improve
maintainability and make it easier for LLMs to work with the codebase:

- grpc_server.go (92 lines): Core type and helper functions
- grpc_server_window.go (243): Window operations
- grpc_server_widgets_basic.go (337): Basic widgets (Label, Button, etc.)
- grpc_server_containers.go (962): Container creation and operations
- grpc_server_display.go (213): Display widgets (Separator, ProgressBar, etc.)
- grpc_server_inputs.go (164): Input widgets (Slider, RadioGroup, etc.)
- grpc_server_data.go (502): Data widgets and operations (Table, List, etc.)
- grpc_server_canvas.go (654): Canvas primitives, updates, and desktop widgets
- grpc_server_operations.go (759): Widget operations and event handling
- grpc_server_state.go (408): Widget state and interaction operations
- grpc_server_dialogs.go (457): Dialog operations
- grpc_server_theme.go (568): Theme, styling, platform, and accessibility

No functional changes - the decomposition preserves all gRPC API behavior.
