+++ bridge/grpc_server_canvas.go
@@ -1,0 +1,654 @@
+package main
+
+import (
+	"context"
+	"encoding/base64"
+
+	pb "github.com/paul-hammant/tsyne/bridge/proto"
+)
+
+// ============================================================================
+// Canvas primitives
+// ============================================================================
+
+// CreateCanvasLine creates a canvas line
+func (s *grpcBridgeService) CreateCanvasLine(ctx context.Context, req *pb.CreateCanvasLineRequest) (*pb.Response, error) {
+	msg := Message{
+		ID:   req.WidgetId,
+		Type: "createCanvasLine",
+		Payload: map[string]interface{}{
+			"id":          req.WidgetId,
+			"x1":          float64(req.X1),
+			"y1":          float64(req.Y1),
+			"x2":          float64(req.X2),
+			"y2":          float64(req.Y2),
+			"strokeColor": req.StrokeColor,
+			"strokeWidth": float64(req.StrokeWidth),
+		},
+	}
+
+	resp := s.bridge.handleCreateCanvasLine(msg)
+
+	return &pb.Response{
+		Success: resp.Success,
+		Error:   resp.Error,
+	}, nil
+}
+
+// CreateCanvasCircle creates a canvas circle
+func (s *grpcBridgeService) CreateCanvasCircle(ctx context.Context, req *pb.CreateCanvasCircleRequest) (*pb.Response, error) {
+	msg := Message{
+		ID:   req.WidgetId,
+		Type: "createCanvasCircle",
+		Payload: map[string]interface{}{
+			"id":          req.WidgetId,
+			"x":           float64(req.X),
+			"y":           float64(req.Y),
+			"x2":          float64(req.X2),
+			"y2":          float64(req.Y2),
+			"fillColor":   req.FillColor,
+			"strokeColor": req.StrokeColor,
+			"strokeWidth": float64(req.StrokeWidth),
+		},
+	}
+
+	resp := s.bridge.handleCreateCanvasCircle(msg)
+
+	return &pb.Response{
+		Success: resp.Success,
+		Error:   resp.Error,
+	}, nil
+}
+
+// CreateCanvasRectangle creates a canvas rectangle
+func (s *grpcBridgeService) CreateCanvasRectangle(ctx context.Context, req *pb.CreateCanvasRectangleRequest) (*pb.Response, error) {
+	payload := map[string]interface{}{
+		"id":          req.WidgetId,
+		"width":       float64(req.Width),
+		"height":      float64(req.Height),
+		"fillColor":   req.FillColor,
+		"strokeColor": req.StrokeColor,
+		"strokeWidth": float64(req.StrokeWidth),
+	}
+	if req.CornerRadius != 0 {
+		payload["cornerRadius"] = float64(req.CornerRadius)
+	}
+
+	msg := Message{
+		ID:      req.WidgetId,
+		Type:    "createCanvasRectangle",
+		Payload: payload,
+	}
+
+	resp := s.bridge.handleCreateCanvasRectangle(msg)
+
+	return &pb.Response{
+		Success: resp.Success,
+		Error:   resp.Error,
+	}, nil
+}
+
+// CreateCanvasText creates canvas text
+func (s *grpcBridgeService) CreateCanvasText(ctx context.Context, req *pb.CreateCanvasTextRequest) (*pb.Response, error) {
+	payload := map[string]interface{}{
+		"id":    req.WidgetId,
+		"text":  req.Text,
+		"color": req.Color,
+	}
+	if req.TextSize != 0 {
+		payload["textSize"] = float64(req.TextSize)
+	}
+	if req.Bold {
+		payload["bold"] = true
+	}
+	if req.Italic {
+		payload["italic"] = true
+	}
+	if req.Monospace {
+		payload["monospace"] = true
+	}
+	if req.Alignment != pb.TextAlignment_TEXT_ALIGN_LEADING {
+		switch req.Alignment {
+		case pb.TextAlignment_TEXT_ALIGN_CENTER:
+			payload["alignment"] = "center"
+		case pb.TextAlignment_TEXT_ALIGN_TRAILING:
+			payload["alignment"] = "trailing"
+		}
+	}
+
+	msg := Message{
+		ID:      req.WidgetId,
+		Type:    "createCanvasText",
+		Payload: payload,
+	}
+
+	resp := s.bridge.handleCreateCanvasText(msg)
+
+	return &pb.Response{
+		Success: resp.Success,
+		Error:   resp.Error,
+	}, nil
+}
+
+// CreateCanvasRaster creates a canvas raster
+func (s *grpcBridgeService) CreateCanvasRaster(ctx context.Context, req *pb.CreateCanvasRasterRequest) (*pb.Response, error) {
+	// Convert pixels to base64
+	pixelData := base64.StdEncoding.EncodeToString(req.Pixels)
+
+	msg := Message{
+		ID:   req.WidgetId,
+		Type: "createCanvasRaster",
+		Payload: map[string]interface{}{
+			"id":     req.WidgetId,
+			"width":  float64(req.Width),
+			"height": float64(req.Height),
+			"pixels": pixelData,
+		},
+	}
+
+	resp := s.bridge.handleCreateCanvasRaster(msg)
+
+	return &pb.Response{
+		Success: resp.Success,
+		Error:   resp.Error,
+	}, nil
+}
+
+// CreateCanvasLinearGradient creates a linear gradient
+func (s *grpcBridgeService) CreateCanvasLinearGradient(ctx context.Context, req *pb.CreateCanvasLinearGradientRequest) (*pb.Response, error) {
+	msg := Message{
+		ID:   req.WidgetId,
+		Type: "createCanvasLinearGradient",
+		Payload: map[string]interface{}{
+			"id":         req.WidgetId,
+			"startColor": req.StartColor,
+			"endColor":   req.EndColor,
+			"angle":      req.Angle,
+			"width":      float64(req.Width),
+			"height":     float64(req.Height),
+		},
+	}
+
+	resp := s.bridge.handleCreateCanvasLinearGradient(msg)
+
+	return &pb.Response{
+		Success: resp.Success,
+		Error:   resp.Error,
+	}, nil
+}
+
+// CreateCanvasRadialGradient creates a radial gradient
+func (s *grpcBridgeService) CreateCanvasRadialGradient(ctx context.Context, req *pb.CreateCanvasRadialGradientRequest) (*pb.Response, error) {
+	msg := Message{
+		ID:   req.WidgetId,
+		Type: "createCanvasRadialGradient",
+		Payload: map[string]interface{}{
+			"id":            req.WidgetId,
+			"startColor":    req.StartColor,
+			"endColor":      req.EndColor,
+			"centerOffsetX": req.CenterOffsetX,
+			"centerOffsetY": req.CenterOffsetY,
+			"width":         float64(req.Width),
+			"height":        float64(req.Height),
+		},
+	}
+
+	resp := s.bridge.handleCreateCanvasRadialGradient(msg)
+
+	return &pb.Response{
+		Success: resp.Success,
+		Error:   resp.Error,
+	}, nil
+}
+
+// CreateCanvasArc creates a canvas arc
+func (s *grpcBridgeService) CreateCanvasArc(ctx context.Context, req *pb.CreateCanvasArcRequest) (*pb.Response, error) {
+	msg := Message{
+		ID:   req.WidgetId,
+		Type: "createCanvasArc",
+		Payload: map[string]interface{}{
+			"id":          req.WidgetId,
+			"x":           float64(req.X),
+			"y":           float64(req.Y),
+			"x2":          float64(req.X2),
+			"y2":          float64(req.Y2),
+			"startAngle":  req.StartAngle,
+			"endAngle":    req.EndAngle,
+			"fillColor":   req.FillColor,
+			"strokeColor": req.StrokeColor,
+			"strokeWidth": float64(req.StrokeWidth),
+		},
+	}
+
+	resp := s.bridge.handleCreateCanvasArc(msg)
+
+	return &pb.Response{
+		Success: resp.Success,
+		Error:   resp.Error,
+	}, nil
+}
+
+// CreateCanvasPolygon creates a canvas polygon
+func (s *grpcBridgeService) CreateCanvasPolygon(ctx context.Context, req *pb.CreateCanvasPolygonRequest) (*pb.Response, error) {
+	points := make([]interface{}, len(req.Points))
+	for i, pt := range req.Points {
+		points[i] = map[string]interface{}{
+			"x": float64(pt.X),
+			"y": float64(pt.Y),
+		}
+	}
+
+	msg := Message{
+		ID:   req.WidgetId,
+		Type: "createCanvasPolygon",
+		Payload: map[string]interface{}{
+			"id":          req.WidgetId,
+			"points":      points,
+			"fillColor":   req.FillColor,
+			"strokeColor": req.StrokeColor,
+			"strokeWidth": float64(req.StrokeWidth),
+		},
+	}
+
+	resp := s.bridge.handleCreateCanvasPolygon(msg)
+
+	return &pb.Response{
+		Success: resp.Success,
+		Error:   resp.Error,
+	}, nil
+}
+
+// CreateTappableCanvasRaster creates a tappable canvas raster
+func (s *grpcBridgeService) CreateTappableCanvasRaster(ctx context.Context, req *pb.CreateTappableCanvasRasterRequest) (*pb.Response, error) {
+	pixelData := base64.StdEncoding.EncodeToString(req.Pixels)
+
+	msg := Message{
+		ID:   req.WidgetId,
+		Type: "createTappableCanvasRaster",
+		Payload: map[string]interface{}{
+			"id":     req.WidgetId,
+			"width":  float64(req.Width),
+			"height": float64(req.Height),
+			"pixels": pixelData,
+		},
+	}
+
+	resp := s.bridge.handleCreateTappableCanvasRaster(msg)
+
+	return &pb.Response{
+		Success: resp.Success,
+		Error:   resp.Error,
+	}, nil
+}
+
+// ============================================================================
+// Canvas updates
+// ============================================================================
+
+// UpdateCanvasLine updates a canvas line
+func (s *grpcBridgeService) UpdateCanvasLine(ctx context.Context, req *pb.UpdateCanvasLineRequest) (*pb.Response, error) {
+	msg := Message{
+		ID:   req.WidgetId,
+		Type: "updateCanvasLine",
+		Payload: map[string]interface{}{
+			"widgetId":    req.WidgetId,
+			"x1":          float64(req.X1),
+			"y1":          float64(req.Y1),
+			"x2":          float64(req.X2),
+			"y2":          float64(req.Y2),
+			"strokeColor": req.StrokeColor,
+			"strokeWidth": float64(req.StrokeWidth),
+		},
+	}
+
+	resp := s.bridge.handleUpdateCanvasLine(msg)
+
+	return &pb.Response{
+		Success: resp.Success,
+		Error:   resp.Error,
+	}, nil
+}
+
+// UpdateCanvasCircle updates a canvas circle
+func (s *grpcBridgeService) UpdateCanvasCircle(ctx context.Context, req *pb.UpdateCanvasCircleRequest) (*pb.Response, error) {
+	msg := Message{
+		ID:   req.WidgetId,
+		Type: "updateCanvasCircle",
+		Payload: map[string]interface{}{
+			"widgetId":    req.WidgetId,
+			"x":           float64(req.X),
+			"y":           float64(req.Y),
+			"x2":          float64(req.X2),
+			"y2":          float64(req.Y2),
+			"fillColor":   req.FillColor,
+			"strokeColor": req.StrokeColor,
+			"strokeWidth": float64(req.StrokeWidth),
+		},
+	}
+
+	resp := s.bridge.handleUpdateCanvasCircle(msg)
+
+	return &pb.Response{
+		Success: resp.Success,
+		Error:   resp.Error,
+	}, nil
+}
+
+// UpdateCanvasRectangle updates a canvas rectangle
+func (s *grpcBridgeService) UpdateCanvasRectangle(ctx context.Context, req *pb.UpdateCanvasRectangleRequest) (*pb.Response, error) {
+	payload := map[string]interface{}{
+		"widgetId":    req.WidgetId,
+		"width":       float64(req.Width),
+		"height":      float64(req.Height),
+		"fillColor":   req.FillColor,
+		"strokeColor": req.StrokeColor,
+		"strokeWidth": float64(req.StrokeWidth),
+	}
+	if req.CornerRadius != 0 {
+		payload["cornerRadius"] = float64(req.CornerRadius)
+	}
+
+	msg := Message{
+		ID:      req.WidgetId,
+		Type:    "updateCanvasRectangle",
+		Payload: payload,
+	}
+
+	resp := s.bridge.handleUpdateCanvasRectangle(msg)
+
+	return &pb.Response{
+		Success: resp.Success,
+		Error:   resp.Error,
+	}, nil
+}
+
+// UpdateCanvasText updates canvas text
+func (s *grpcBridgeService) UpdateCanvasText(ctx context.Context, req *pb.UpdateCanvasTextRequest) (*pb.Response, error) {
+	payload := map[string]interface{}{
+		"widgetId": req.WidgetId,
+		"text":     req.Text,
+		"color":    req.Color,
+	}
+	if req.TextSize != 0 {
+		payload["textSize"] = float64(req.TextSize)
+	}
+
+	msg := Message{
+		ID:      req.WidgetId,
+		Type:    "updateCanvasText",
+		Payload: payload,
+	}
+
+	resp := s.bridge.handleUpdateCanvasText(msg)
+
+	return &pb.Response{
+		Success: resp.Success,
+		Error:   resp.Error,
+	}, nil
+}
+
+// UpdateCanvasRaster updates a canvas raster
+func (s *grpcBridgeService) UpdateCanvasRaster(ctx context.Context, req *pb.UpdateCanvasRasterRequest) (*pb.Response, error) {
+	updates := make([]interface{}, len(req.Updates))
+	for i, u := range req.Updates {
+		updates[i] = map[string]interface{}{
+			"x": float64(u.X),
+			"y": float64(u.Y),
+			"r": float64(u.R),
+			"g": float64(u.G),
+			"b": float64(u.B),
+			"a": float64(u.A),
+		}
+	}
+
+	msg := Message{
+		ID:   req.WidgetId,
+		Type: "updateCanvasRaster",
+		Payload: map[string]interface{}{
+			"widgetId": req.WidgetId,
+			"updates":  updates,
+		},
+	}
+
+	resp := s.bridge.handleUpdateCanvasRaster(msg)
+
+	return &pb.Response{
+		Success: resp.Success,
+		Error:   resp.Error,
+	}, nil
+}
+
+// UpdateCanvasLinearGradient updates a linear gradient
+func (s *grpcBridgeService) UpdateCanvasLinearGradient(ctx context.Context, req *pb.UpdateCanvasLinearGradientRequest) (*pb.Response, error) {
+	msg := Message{
+		ID:   req.WidgetId,
+		Type: "updateCanvasLinearGradient",
+		Payload: map[string]interface{}{
+			"widgetId":   req.WidgetId,
+			"startColor": req.StartColor,
+			"endColor":   req.EndColor,
+			"angle":      req.Angle,
+		},
+	}
+
+	resp := s.bridge.handleUpdateCanvasLinearGradient(msg)
+
+	return &pb.Response{
+		Success: resp.Success,
+		Error:   resp.Error,
+	}, nil
+}
+
+// UpdateCanvasRadialGradient updates a radial gradient
+func (s *grpcBridgeService) UpdateCanvasRadialGradient(ctx context.Context, req *pb.UpdateCanvasRadialGradientRequest) (*pb.Response, error) {
+	msg := Message{
+		ID:   req.WidgetId,
+		Type: "updateCanvasRadialGradient",
+		Payload: map[string]interface{}{
+			"widgetId":      req.WidgetId,
+			"startColor":    req.StartColor,
+			"endColor":      req.EndColor,
+			"centerOffsetX": req.CenterOffsetX,
+			"centerOffsetY": req.CenterOffsetY,
+		},
+	}
+
+	resp := s.bridge.handleUpdateCanvasRadialGradient(msg)
+
+	return &pb.Response{
+		Success: resp.Success,
+		Error:   resp.Error,
+	}, nil
+}
+
+// UpdateCanvasArc updates a canvas arc
+func (s *grpcBridgeService) UpdateCanvasArc(ctx context.Context, req *pb.UpdateCanvasArcRequest) (*pb.Response, error) {
+	msg := Message{
+		ID:   req.WidgetId,
+		Type: "updateCanvasArc",
+		Payload: map[string]interface{}{
+			"widgetId":    req.WidgetId,
+			"startAngle":  req.StartAngle,
+			"endAngle":    req.EndAngle,
+			"fillColor":   req.FillColor,
+			"strokeColor": req.StrokeColor,
+			"strokeWidth": float64(req.StrokeWidth),
+		},
+	}
+
+	resp := s.bridge.handleUpdateCanvasArc(msg)
+
+	return &pb.Response{
+		Success: resp.Success,
+		Error:   resp.Error,
+	}, nil
+}
+
+// UpdateCanvasPolygon updates a canvas polygon
+func (s *grpcBridgeService) UpdateCanvasPolygon(ctx context.Context, req *pb.UpdateCanvasPolygonRequest) (*pb.Response, error) {
+	points := make([]interface{}, len(req.Points))
+	for i, pt := range req.Points {
+		points[i] = map[string]interface{}{
+			"x": float64(pt.X),
+			"y": float64(pt.Y),
+		}
+	}
+
+	msg := Message{
+		ID:   req.WidgetId,
+		Type: "updateCanvasPolygon",
+		Payload: map[string]interface{}{
+			"widgetId":    req.WidgetId,
+			"points":      points,
+			"fillColor":   req.FillColor,
+			"strokeColor": req.StrokeColor,
+			"strokeWidth": float64(req.StrokeWidth),
+		},
+	}
+
+	resp := s.bridge.handleUpdateCanvasPolygon(msg)
+
+	return &pb.Response{
+		Success: resp.Success,
+		Error:   resp.Error,
+	}, nil
+}
+
+// UpdateTappableCanvasRaster updates a tappable canvas raster
+func (s *grpcBridgeService) UpdateTappableCanvasRaster(ctx context.Context, req *pb.UpdateTappableCanvasRasterRequest) (*pb.Response, error) {
+	updates := make([]interface{}, len(req.Updates))
+	for i, u := range req.Updates {
+		updates[i] = map[string]interface{}{
+			"x": float64(u.X),
+			"y": float64(u.Y),
+			"r": float64(u.R),
+			"g": float64(u.G),
+			"b": float64(u.B),
+			"a": float64(u.A),
+		}
+	}
+
+	msg := Message{
+		ID:   req.WidgetId,
+		Type: "updateTappableCanvasRaster",
+		Payload: map[string]interface{}{
+			"widgetId": req.WidgetId,
+			"updates":  updates,
+		},
+	}
+
+	resp := s.bridge.handleUpdateTappableCanvasRaster(msg)
+
+	return &pb.Response{
+		Success: resp.Success,
+		Error:   resp.Error,
+	}, nil
+}
+
+// ============================================================================
+// Desktop widgets
+// ============================================================================
+
+// CreateDesktopCanvas creates a desktop canvas widget
+func (s *grpcBridgeService) CreateDesktopCanvas(ctx context.Context, req *pb.CreateDesktopCanvasRequest) (*pb.Response, error) {
+	msg := Message{
+		ID:   req.WidgetId,
+		Type: "createDesktopCanvas",
+		Payload: map[string]interface{}{
+			"id":      req.WidgetId,
+			"bgColor": req.BgColor,
+		},
+	}
+
+	resp := s.bridge.handleCreateDesktopCanvas(msg)
+
+	return &pb.Response{
+		Success: resp.Success,
+		Error:   resp.Error,
+	}, nil
+}
+
+// CreateDesktopIcon creates a desktop icon
+func (s *grpcBridgeService) CreateDesktopIcon(ctx context.Context, req *pb.CreateDesktopIconRequest) (*pb.Response, error) {
+	msg := Message{
+		ID:   req.WidgetId,
+		Type: "createDesktopIcon",
+		Payload: map[string]interface{}{
+			"id":                   req.WidgetId,
+			"desktopId":            req.DesktopId,
+			"label":                req.Label,
+			"x":                    float64(req.X),
+			"y":                    float64(req.Y),
+			"color":                req.Color,
+			"onClickCallbackId":    req.OnClickCallbackId,
+			"onDblClickCallbackId": req.OnDblClickCallbackId,
+			"onDragCallbackId":     req.DragCallbackId,
+			"onDragEndCallbackId":  req.DragEndCallbackId,
+		},
+	}
+
+	resp := s.bridge.handleCreateDesktopIcon(msg)
+
+	return &pb.Response{
+		Success: resp.Success,
+		Error:   resp.Error,
+	}, nil
+}
+
+// MoveDesktopIcon moves a desktop icon
+func (s *grpcBridgeService) MoveDesktopIcon(ctx context.Context, req *pb.MoveDesktopIconRequest) (*pb.Response, error) {
+	msg := Message{
+		ID:   req.IconId,
+		Type: "moveDesktopIcon",
+		Payload: map[string]interface{}{
+			"iconId": req.IconId,
+			"x":      float64(req.X),
+			"y":      float64(req.Y),
+		},
+	}
+
+	resp := s.bridge.handleMoveDesktopIcon(msg)
+
+	return &pb.Response{
+		Success: resp.Success,
+		Error:   resp.Error,
+	}, nil
+}
+
+// UpdateDesktopIconLabel updates a desktop icon's label
+func (s *grpcBridgeService) UpdateDesktopIconLabel(ctx context.Context, req *pb.UpdateDesktopIconLabelRequest) (*pb.Response, error) {
+	msg := Message{
+		ID:   req.IconId,
+		Type: "updateDesktopIconLabel",
+		Payload: map[string]interface{}{
+			"iconId": req.IconId,
+			"label":  req.Label,
+		},
+	}
+
+	resp := s.bridge.handleUpdateDesktopIconLabel(msg)
+
+	return &pb.Response{
+		Success: resp.Success,
+		Error:   resp.Error,
+	}, nil
+}
+
+// UpdateDesktopIconColor updates a desktop icon's color
+func (s *grpcBridgeService) UpdateDesktopIconColor(ctx context.Context, req *pb.UpdateDesktopIconColorRequest) (*pb.Response, error) {
+	msg := Message{
+		ID:   req.IconId,
+		Type: "updateDesktopIconColor",
+		Payload: map[string]interface{}{
+			"iconId": req.IconId,
+			"color":  req.Color,
+		},
+	}
+
+	resp := s.bridge.handleUpdateDesktopIconColor(msg)
+
+	return &pb.Response{
+		Success: resp.Success,
+		Error:   resp.Error,
+	}, nil
+}
